requirements.txt
streamlit>=1.30
pillow>=10.0.0
numpy>=1.24
pandas>=2.0
opencv-python-headless>=4.8
PyMuPDF>=1.23
python-dateutil>=2.8
rapidocr-onnxruntime>=1.4.4
pytest>=7.0
